% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{commandCenterCreater}
\alias{commandCenterCreater}
\title{create a commandcenter file. This is the function you would run to create the new dashboard}
\usage{
commandCenterCreater(
  pathToRuns = "..",
  hrs_lag = 24,
  outputFolders = "outputs",
  ccFolder = "../cc",
  minute = "15",
  hour = "*",
  dayOfMonth = "*",
  month = "*",
  dayOfWeek = "*"
)
}
\arguments{
\item{pathToRuns}{path to all the runner folders, Default: '..'}

\item{hrs_lag}{how many hours ago, Default: 24}

\item{outputFolders}{what do you name your output folders?, Default: 'Outputs'}

\item{ccFolder}{what is the name of the folder for your commandcenter?, Default: 'cc'}

\item{minute}{in which minute(s) should the script run? 0-59, or \* for every minute, or
comma seperated list like 1,10,20 means minute 1, 10, 20, or -    range of values like
1-9 means 1,2,3,4,5,6,7,8,9, or /    step values like \*/15 means every fifteen minutes. Default 15,
which means that it'll run once an hour (on minute 15 of every hour of every day of every...)}

\item{hour}{in which hour(s) should the script run? 0-23, and "\*", ",", "-", "/" accepted}

\item{dayOfMonth}{in what days of the month should run? 1-31, , and "\*", ",", "-", "/" accepted}

\item{month}{in what months should the script run? 1-12, and "\*", ",", "-", "/" accepted}

\item{dayOfWeek}{in what days of the week should the script run? 0-6 (starting sunday),
and "\*", ",", "-", "/" accepted}
}
\value{
outputs a file in the commandcenter
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[rmarkdown]{render}}
}
