% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{createAssets}
\alias{createAssets}
\title{automation creater}
\usage{
createAssets(
  file,
  ultimatePath = getwd(),
  commandCenterPath = getwd(),
  minute = "1",
  hour = "*",
  dayOfMonth = "*",
  month = "*",
  dayOfWeek = "*"
)
}
\arguments{
\item{file}{which R file should be automated}

\item{ultimatePath}{what is the FULL final path containing the file, Default: getwd()}

\item{commandCenterPath}{if you'd the timeRun logs to go to another central folder,
place FULL path here , Default: getwd()}

\item{minute}{in which minute(s) should the script run? 0-59, or \* for every minute, or
comma seperated list like 1,10,20 means minute 1, 10, 20, or -    range of values like
1-9 means 1,2,3,4,5,6,7,8,9, or /    step values like \*/15 means every fifteen minutes. Default 1,
which means that it'll run once an hour (on minute 1 of every hour of every day of every...)}

\item{hour}{in which hour(s) should the script run? 0-23, and "\*", ",", "-", "/" accepted}

\item{dayOfMonth}{in what days of the month should run? 1-31, , and "\*", ",", "-", "/" accepted}

\item{month}{in what months should the script run? 1-12, and "\*", ",", "-", "/" accepted}

\item{dayOfWeek}{in what days of the week should the script run? 0-6 (starting sunday),
and "\*", ",", "-", "/" accepted}
}
\value{
nothing is returned, but the crontab entry and other info is outputted to the console and files being created.
}
\description{
this function creates the scripts required to automate a script
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 cat("", file = "example.R")
 createAssets("example.R", getwd(), getwd(), 1,2,3,4,5)
 file.remove("example.R")

 }
}
}
\seealso{
\code{\link[amitFuncs]{right}}
}
