% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{commandCenterCreater}
\alias{commandCenterCreater}
\title{create a commandcenter file. This is the function you would run to create the new dashboard}
\usage{
commandCenterCreater(
  pathToRuns = "..",
  hrs = 24,
  outputFolders = "outputs",
  ccFolder = "../cc"
)
}
\arguments{
\item{pathToRuns}{path to all the runner folders, Default: '..'}

\item{hrs}{how many hours ago, Default: 24}

\item{outputFolders}{what do you name your output folders?, Default: 'Outputs'}

\item{ccFolder}{what is the name of the folder for your commandcenter?, Default: 'cc'}
}
\value{
outputs a file in the commandcenter
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[rmarkdown]{render}}
}
